made remove jobs modular
