+++ extension/manifest.json
@@ -1,7 +1,7 @@
 {
   "manifest_version": 3,
   "name": "Highlight stop-words on Indeed page",
-  "version": "0.4",
+  "version": "0.41",
   "description": "If Indeed's container has restricted word secret or clearance - heighlight it and change background of root container",
   "icons": {
     "16": "images/icon-16.png",
