+++ extension/scripts/content.js
@@ -1,1 +1,1 @@
-(()=>{"use strict";const e={HIGHLIGHT_WORDS:"stopWords",REMOVE_WORDS:"removeWords",COLORS:"colors",SELECTORS:"selectors"},o={[e.SELECTORS]:{SCROLLABLE_CONTAINER:"jobsearch-JobComponent",SKELETON_CLASS:"jobsearch-ViewJobSkeleton",REMOVE_TEST_ID:"viewJob-skeleton"},[e.COLORS]:{highlightColor:"#f3d381",backgroundColor:"#b0b0b0"},[e.REMOVE_WORDS]:["senior","lead","architect","java",".net","sr.","manager"],[e.HIGHLIGHT_WORDS]:["secret","clearance","citizen"]},{HIGHLIGHT_WORDS:t,REMOVE_WORDS:r,COLORS:n,SELECTORS:s}=e,c=Object.assign({},o);chrome.storage.sync.get([t,r,n,s],(e=>{chrome.runtime.lastError?console.error(chrome.runtime.lastError):(console.log(e),c[t]=e[t],c[r]=e[r],c[n]=e[n],c[s]=e[s])}));const l=()=>{var e;const o=document.getElementsByClassName(c[s].SCROLLABLE_CONTAINER)[0];if(!o)return;Array.isArray(c[t])&&(null===(e=c[t])||void 0===e||e.forEach((e=>function(e,o,t){if(!(o instanceof Node))return;const r=document.createTreeWalker(o,NodeFilter.SHOW_TEXT);let n,s=!1;const c=new RegExp(e,"gi");for(;n=r.nextNode();)(null==n?void 0:n.textContent)&&n.textContent.match(c)&&(s=!0,n.textContent=n.textContent.replace(c,`@@@${e}@@@`));s&&(o.innerHTML=o.innerHTML.replace(/@@@(.*?)@@@/g,`<span style="background-color: ${t.highlightColor}">$1</span>`),o.style.backgroundColor=t.backgroundColor)}(e,o,c[n]))));const r=o.querySelector(".jobsearch-JobComponent-embeddedBody");if(r){const e=r.querySelector(`span[style*="background-color: ${c[n].highlightColor}"]`);e&&(r.scrollTop=e.offsetTop-260)}else console.error("check selector '.jobsearch-JobComponent-embeddedBody'")},a=()=>{const e=c[r].map((e=>RegExp.escape(e))).join("|"),o=new RegExp(`\\b(${e})\\b`,"gi"),t=Array.from(document.querySelectorAll("h2.jobTitle"));if(0===t.length)return;const n=t.filter((e=>o.test(e.textContent)));0!==n.length&&n.forEach((e=>{const o=e.closest("li");o&&o.remove()}))};new MutationObserver((e=>{for(let o of e)"childList"===o.type&&o.removedNodes.forEach((e=>{console.log(e),e.nodeType!==Node.ELEMENT_NODE||!e.classList.contains(c[s].SKELETON_CLASS)&&e.getAttribute("data-testid")!==c[s].REMOVE_TEST_ID||(console.log("Skeleton removed!"),l(),a())}))})).observe(document,{childList:!0,subtree:!0})})();
+(()=>{"use strict";const e={HIGHLIGHT_WORDS:"stopWords",REMOVE_WORDS:"removeWords",COLORS:"colors",SELECTORS:"selectors"},o={[e.SELECTORS]:{SCROLLABLE_CONTAINER:"jobsearch-JobComponent",SKELETON_CLASS:"jobsearch-ViewJobSkeleton",REMOVE_TEST_ID:"viewJob-skeleton"},[e.COLORS]:{highlightColor:"#f3d381",backgroundColor:"#b0b0b0"},[e.REMOVE_WORDS]:["senior","lead","architect","java",".net","sr.","manager"],[e.HIGHLIGHT_WORDS]:["secret","clearance","citizen"]},t=e=>{const o=function(e){return e.map((e=>e.replace(/[.*+?^${}()|[\]\\]/g,"\\$&"))).join("|")}(e),t=new RegExp(`\\b(${o})\\b`,"gi"),r=Array.from(document.querySelectorAll("h2.jobTitle"));if(0===r.length)return;const n=r.filter((e=>t.test(e.textContent)));0!==n.length&&n.forEach((e=>{const o=e.closest("li");o&&o.remove()}))},{HIGHLIGHT_WORDS:r,REMOVE_WORDS:n,COLORS:c,SELECTORS:s}=e,l=Object.assign({},o);chrome.storage.sync.get([r,n,c,s],(e=>{chrome.runtime.lastError?console.error(chrome.runtime.lastError):(console.log(e),l[r]=e[r],l[n]=e[n],l[c]=e[c],l[s]=e[s])}));const a=()=>{var e;const o=document.getElementsByClassName(l[s].SCROLLABLE_CONTAINER)[0];if(!o)return;Array.isArray(l[r])&&(null===(e=l[r])||void 0===e||e.forEach((e=>function(e,o,t){if(!(o instanceof Node))return;const r=document.createTreeWalker(o,NodeFilter.SHOW_TEXT);let n,c=!1;const s=new RegExp(e,"gi");for(;n=r.nextNode();)(null==n?void 0:n.textContent)&&n.textContent.match(s)&&(c=!0,n.textContent=n.textContent.replace(s,`@@@${e}@@@`));c&&(o.innerHTML=o.innerHTML.replace(/@@@(.*?)@@@/g,`<span style="background-color: ${t.highlightColor}">$1</span>`),o.style.backgroundColor=t.backgroundColor)}(e,o,l[c]))));const t=o.querySelector(".jobsearch-JobComponent-embeddedBody");if(t){const e=t.querySelector(`span[style*="background-color: ${l[c].highlightColor}"]`);e&&(t.scrollTop=e.offsetTop-260)}else console.error("check selector '.jobsearch-JobComponent-embeddedBody'")};new MutationObserver((e=>{for(let o of e)"childList"===o.type&&o.removedNodes.forEach((e=>{console.log(e),e.nodeType!==Node.ELEMENT_NODE||!e.classList.contains(l[s].SKELETON_CLASS)&&e.getAttribute("data-testid")!==l[s].REMOVE_TEST_ID||(console.log("Skeleton removed!"),a(),t(l[n]))}))})).observe(document,{childList:!0,subtree:!0})})();
